Work on non-single level commands
