+++ Samples/StarFruit.Samples/StarFruit.Samples.csproj
@@ -9,8 +9,8 @@
     <ProjectReference Include="..\..\StarFruit2\StarFruit2.csproj" />
   </ItemGroup>
   
-  <ItemGroup>
+  <!--<ItemGroup>
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20371.2" />
-  </ItemGroup>
+  </ItemGroup>-->
 
 </Project>
