+++ StarFruit2/StarFruit2.csproj
@@ -8,11 +8,11 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.7.0" />
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20371.2" />
+    <!--<PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20371.2" />-->
   </ItemGroup>
 
   <ItemGroup>
 		<ProjectReference Include="..\StarFruit.Common\StarFruit.Common.csproj" />
-	  </ItemGroup>
+	</ItemGroup>
 
 	</Project>
