+++ Starfruit2_B/Starfruit2_B.csproj
@@ -8,7 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.7.0" />
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20371.2" />
+    <!--<PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20371.2" />-->
   </ItemGroup>
 
   <ItemGroup>
